source_code_v1/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
attrs==25.3.0
audioread==3.0.1
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
colorama==0.4.6
datasets==2.16.1
decorator==5.2.1
dill==0.3.7
distro==1.9.0
exceptiongroup==1.3.0
filelock==3.18.0
frozenlist==1.7.0
fsspec==2023.10.0
h11==0.16.0
httpcore==1.0.9
httpx==0.27.2
huggingface-hub==0.34.1
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
joblib==1.5.1
lazy_loader==0.4
librosa==0.10.1
llvmlite==0.44.0
loguru==0.7.2
MarkupSafe==3.0.2
more-itertools==10.7.0
mpmath==1.3.0
msgpack==1.1.1
multidict==6.6.3
multiprocess==0.70.15
networkx==3.4.2
numba==0.61.2
numpy==1.24.3
openai==1.12.0
openai-whisper==20231117
packaging==25.0
pandas==2.0.3
platformdirs==4.3.8
pluggy==1.6.0
pooch==1.8.2
propcache==0.3.2
pyarrow==21.0.0
pyarrow-hotfix==0.7
# PyAudio==0.2.11
pycparser==2.22
pydantic==2.5.3
pydantic_core==2.14.6
pytest==7.4.4
pytest-asyncio==0.23.2
pytest-mock==3.12.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2025.2
PyYAML==6.0.1
regex==2024.11.6
requests==2.32.4
safetensors==0.5.3
scikit-learn==1.7.1
scipy>=1.7.0
six==1.17.0
sniffio==1.3.1
sounddevice>=0.4.0
soundfile==0.12.1
soxr==0.5.0.post1
sympy==1.14.0
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.15.2
tomli==2.2.1
torch>=1.9.0
torchaudio==2.1.2
tqdm==4.67.1
transformers==4.37.2
typing_extensions==4.9.0
tzdata==2025.2
urllib3==2.5.0
win32_setctime==1.2.0
xxhash==3.5.0
yarl==1.20.1

# FastAPI 관련 의존성
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# 보안 관련 의존성
python-magic==0.4.27

# 고급 화자 분리를 위한 추가 의존성
pyannote.audio==3.3.2
speechbrain==1.0.3
